fix frozen hands and try to fix walking speed bug

+++ _public_server/PlayerStats.cs
@@ -769,7 +769,7 @@ public class PlayerStats : NetworkBehaviour
 
     #region Mods
     //Normales        
-    public int modSTR,
+    public float modSTR,
      modDEX,
      modINT,
      modWIS,
